+++ Simple3DViewer/src/main/java/com/cgvsu/rasterization/Rasterization.java
@@ -18,7 +18,7 @@ public class Rasterization {
                                     final int tx1, final int ty1,
                                     final int tx2, final int ty2,
                                     Color color,
-                                    Double[][] ZBuffer,
+                                    Float[][] ZBuffer,
                                     BufferedImage texture,
                                     boolean useTexture,
                                     boolean useLight,
@@ -33,19 +33,19 @@ public class Rasterization {
         Point2fYThenXComparator compYX = new Point2fYThenXComparator();
         dots.sort(compYX);
 
-        double function1;
-        double function2;
+        float function1;
+        float function2;
         Point2f vertex1 = dots.get(1);
         Point2f vertex2 = dots.get(2);
 
         //Рассчёт функций верхней половины
         if (dots.get(0).y - vertex2.y != 0) {
-            function1 = ((double) (dots.get(0).x - vertex2.x) / (dots.get(0).y - vertex2.y));
+            function1 = ((dots.get(0).x - vertex2.x) / (dots.get(0).y - vertex2.y));
         } else {
             function1 = 0;
         }
         if (dots.get(0).y - vertex1.y != 0) {
-            function2 = (double) (dots.get(0).x - vertex1.x) / (dots.get(0).y - vertex1.y);
+            function2 = (dots.get(0).x - vertex1.x) / (dots.get(0).y - vertex1.y);
         } else {
             function2 = 0;
         }
@@ -57,11 +57,11 @@ public class Rasterization {
         //Рассчёт функций нижней половины
         if (vertex2.y >= vertex1.y) {
             if (dots.get(2).y - vertex1.y != 0) {
-                function2 = (double) (dots.get(2).x - vertex1.x) / (dots.get(2).y - vertex1.y);
+                function2 = (dots.get(2).x - vertex1.x) / (dots.get(2).y - vertex1.y);
             } else function2 = 0;
         } else {
             if (dots.get(2).y - vertex2.y != 0) {
-                function1 = (double) (dots.get(2).x - vertex2.x) / (dots.get(2).y - vertex2.y);
+                function1 = (dots.get(2).x - vertex2.x) / (dots.get(2).y - vertex2.y);
             } else function1 = 0;
         }
 
@@ -71,11 +71,11 @@ public class Rasterization {
         }
     }
 
-    private static void drawHalfOfTriangle(double function1, double function2,
+    private static void drawHalfOfTriangle(float function1, float function2,
                                            Point2f vertex1, Point2f vertex2,
                                            Point2f d0, Point2f d1, Point2f d2,
                                            BufferedImage texture,
-                                           Double[][] ZBuffer,
+                                           Float[][] ZBuffer,
                                            PixelWriter pixelWriter,
                                            int row,
                                            float z0,
@@ -95,17 +95,17 @@ public class Rasterization {
         int fullFunctionLeft = (int) (function1 * row + vertex2.x - function1 * vertex2.y);
         int fullFunctionRight = (int) (function2 * row + vertex1.x - function2 * vertex1.y);
         for (int col = Math.min(fullFunctionRight, fullFunctionLeft); col <= Math.max(fullFunctionRight, fullFunctionLeft); col++) {
-            double[] coordinates = calculateBarycentricCoordinates(col, row, d0, d1, d2);
-            double z = coordinates[0] * z0 + coordinates[1] * z1 + coordinates[2] * z2;
+            float[] coordinates = calculateBarycentricCoordinates(col, row, d0, d1, d2);
+            float z = coordinates[0] * z0 + coordinates[1] * z1 + coordinates[2] * z2;
             if (ZBuffer[col][row] > z) {
                 Color finalColor;
                 if (useTexture) {
                     int x = coordinatesFix((int) (coordinates[0] * tx0 + coordinates[1] * tx1 + coordinates[2] * tx2), texture.getWidth());
                     int y = coordinatesFix((int) (coordinates[0] * ty0 + coordinates[1] * ty1 + coordinates[2] * ty2), texture.getHeight());
                     int clr = texture.getRGB(x, y);
-                    double red = (double) ((clr & 0x00ff0000) >> 16) / 255;
-                    double green = (double) ((clr & 0x0000ff00) >> 8) / 255;
-                    double blue = (double) (clr & 0x000000ff) / 255;
+                    float red = (float) ((clr & 0x00ff0000) >> 16) / 255;
+                    float green = (float) ((clr & 0x0000ff00) >> 8) / 255;
+                    float blue = (float) (clr & 0x000000ff) / 255;
                     finalColor = new Color(red, green, blue, 1);
                 } else {
                     finalColor = color;
@@ -114,8 +114,8 @@ public class Rasterization {
                     Vector3f currentN = new Vector3f(coordinates[0] * n0.getX() + coordinates[1] * n1.getX() + coordinates[2] * n2.getX(),
                             coordinates[0] * n0.getY() + coordinates[1] * n1.getY() + coordinates[2] * n2.getY(),
                             coordinates[0] * n0.getZ() + coordinates[1] * n1.getZ() + coordinates[2] * n2.getZ());
-                    double l = -(currentN.getX()*ray.getX() + currentN.getY()*ray.getY() + currentN.getZ()*ray.getZ());// currentN.vectorMultiply(ray);
-                    double k = 0.7;
+                    float l = -(currentN.getX()*ray.getX() + currentN.getY()*ray.getY() + currentN.getZ()*ray.getZ());// currentN.vectorMultiply(ray);
+                    float k = 0.7f;
                     if (l > 0){
                         if (l>1){
                             l = 1;
@@ -131,25 +131,25 @@ public class Rasterization {
         }
     }
 
-    private static double[] calculateBarycentricCoordinates(int x, int y, Point2f d0, Point2f d1, Point2f d2) {
-        double denominator = (double) (d1.y - d2.y) * (double) (d0.x - d2.x) + (double) (d2.x - d1.x) * (double) (d0.y - d2.y);
-        double alpha;
-        double beta;
-        double gamma;
-        if (Double.isNaN(denominator)) {
+    private static float[] calculateBarycentricCoordinates(int x, int y, Point2f d0, Point2f d1, Point2f d2) {
+        float denominator = (d1.y - d2.y) * (d0.x - d2.x) + (d2.x - d1.x) * (d0.y - d2.y);
+        float alpha;
+        float beta;
+        float gamma;
+        if (Float.isNaN(denominator)) {
             alpha = ((d1.y - d2.y) * (x - d2.x) + (d2.x - d1.x) * (y - d2.y));
             beta = ((d2.y - d0.y) * (x - d2.x) + (d0.x - d2.x) * (y - d2.y));
         } else {
-            alpha = (double) ((d1.y - d2.y) * (x - d2.x) + (d2.x - d1.x) * (y - d2.y)) / denominator;
-            beta = (double) ((d2.y - d0.y) * (x - d2.x) + (d0.x - d2.x) * (y - d2.y)) / denominator;
+            alpha = ((d1.y - d2.y) * (x - d2.x) + (d2.x - d1.x) * (y - d2.y)) / denominator;
+            beta = ((d2.y - d0.y) * (x - d2.x) + (d0.x - d2.x) * (y - d2.y)) / denominator;
         }
         gamma = 1 - alpha - beta;
-        return new double[]{clamp(alpha), clamp(beta), clamp(gamma)};
+        return new float[]{clamp(alpha), clamp(beta), clamp(gamma)};
     }
 
-    private static double clamp(double value) {
-        if (!Double.isNaN(value)) {
-            return Math.max(0.0, Math.min(1.0, value));
+    private static float clamp(float value) {
+        if (!Float.isNaN(value)) {
+            return Math.max(0.0f, Math.min(1.0f, value));
         }
         return 0;
     }
@@ -164,7 +164,7 @@ public class Rasterization {
         return value;
     }
 
-    public static void drawLine(PixelWriter pixelWriter, int x0, int y0, float z0, int x1, int y1, float z1, Double[][] ZBuffer, int width, int height) {
+    public static void drawLine(PixelWriter pixelWriter, int x0, int y0, float z0, int x1, int y1, float z1, Float[][] ZBuffer, int width, int height) {
         int x = x0;
         int y = y0;
         int deltax = Math.abs(x1 - x0);
@@ -174,12 +174,12 @@ public class Rasterization {
         int error = deltax - deltay;
         while (true) {
             if (x0 >= 0 && x0 < width && y0 >= 0 && y0 < height) {
-                double alpha = Math.pow((Math.pow(x - x0, 2) + Math.pow(y - y0, 2)), 0.5);
-                double beta = Math.pow((Math.pow(x1 - x0, 2) + Math.pow(y1 - y0, 2)), 0.5);
-                double denominator = 1 / (alpha + beta);
+                float alpha = (float) Math.pow((Math.pow(x - x0, 2) + Math.pow(y - y0, 2)), 0.5);
+                float beta = (float) Math.pow((Math.pow(x1 - x0, 2) + Math.pow(y1 - y0, 2)), 0.5);
+                float denominator = 1 / (alpha + beta);
                 alpha *= denominator;
                 beta *= denominator;
-                double z = beta * z0 + alpha * z1;
+                float z = beta * z0 + alpha * z1;
                 if (ZBuffer[x0][y0] > z) {
                     pixelWriter.setColor(x0, y0, Color.BLACK);
                     ZBuffer[x0][y0] = z;
